add null generation factory toggle
