+++ packages/plugins/query-generator/src/generator/queryGenerator.ts
@@ -200,6 +200,13 @@ function generateResponse(
     return null
   }
 
+  if (
+    isNullableType(type) &&
+    config.nullGenerationStrategy === NullGenerationStrategy.ALWAYS_NULL
+  ) {
+    return valueResponseBuilder(null)
+  }
+
   if (isLeafType(type)) {
     // No more fields under
     return valueResponseBuilder(
@@ -219,13 +226,6 @@ function generateResponse(
 
   if (isNonNullType(type)) {
     return generateResponse(name, type.ofType, config, context)
-  }
-
-  if (
-    isNullableType(type) &&
-    config.nullGenerationStrategy === NullGenerationStrategy.ALWAYS_NULL
-  ) {
-    return valueResponseBuilder(null)
   }
 
   if (isListType(type)) {
