+++ packages/plugins/query-generator/tests/generator/queryGenerator.spec.ts
@@ -522,8 +522,13 @@ describe('generating a response', () => {
 
       it('generates the response properly', () => {
         const result = generateGraphQLResponse(recursiveField, config)
-
         assertResponseEqual(result, { person: null })
+      })
+
+      it('generates the response properly for a single-level field', () => {
+        const singleLevelField = getQueryField('id')
+        const result = generateGraphQLResponse(singleLevelField, config)
+        assertResponseEqual(result, { id: null })
       })
     })
